fix(logs): rename `code` to `status` on request.summary log lines
Stackdriver doesn't like it when properties change their type across
log lines, and there is a clash between the type of the `code` property
for the `account.signin.confirm.success` and `request.summary` log
lines.

This change renames it to `status` on `request.summary`, because that
seems like a more appropriate name and the other log message appears to
have greater legitimacy to its claim to use the `code` property name
(that is literally the name of the corresponding property in the request
payload).